Enabled Tashkeel support in Arabic
Enabled Tashkeel support in Arabic. This is required for a few select entries in Arabic like "CONTEXT_CONDITION_DEAD".

The second booleon in ArabicFixer.Fix() is "showTashkeel" which are the the vowel marks in Arabic. This is disabled by default, however, I need it to be enabled because I have used that feature in a few entries. My tests show everything works fine when it's enabled.

The third booleon "useHinduNumbers" is maintained as false. I didn't use that anywhere.