Added support for changing font per language
Chinese, Japanese and Korean all use Han characters but some characters are by convention displayed differently per language. This neccesitates using a separate font for each of them. This commit adds a LocalizedFont component which can be added to GameObjects with Text components in order to automatically switch the font based on current language. The font overrides are specified in the Localization asset.